rename the contracts and redeploy
