+++ abis/NftSingle.json
