+++ src/nft-market.ts
@@ -98,21 +98,13 @@ export function handleInitialized(event: InitializedEvent): void {
 }
 
 export function handleMarketItemCreated(event: MarketItemCreatedEvent): void {
-  let nft = NFT.load(
-    Bytes.fromHexString(
-      event.params.nftContract
-        .toHexString()
-        .concat(Bytes.fromBigInt(event.params.tokenId).toHexString())
-    )
+  const nftId = Bytes.fromHexString(
+    event.params.nftContract
+      .toHexString()
+      .concat(Bytes.fromBigInt(event.params.tokenId).toHexString())
   );
-  if (!nft)
-    nft = new NFT(
-      Bytes.fromHexString(
-        event.params.nftContract
-          .toHexString()
-          .concat(Bytes.fromBigInt(event.params.tokenId).toHexString())
-      )
-    );
+  let nft = NFT.load(nftId);
+  if (!nft) nft = new NFT(nftId);
 
   // nft.seller = event.params.seller
   nft.category = event.params.category;
@@ -142,32 +134,36 @@ export function handleMarketItemCreated(event: MarketItemCreatedEvent): void {
 }
 
 export function handleMarketItemSold(event: MarketItemSoldEvent): void {
-  let nft = NFT.load(
-    Bytes.fromHexString(
-      event.params.nftContract
-        .toHexString()
-        .concat(Bytes.fromBigInt(event.params.tokenId).toHexString())
-    )
+  const nftId = Bytes.fromHexString(
+    event.params.nftContract
+      .toHexString()
+      .concat(Bytes.fromBigInt(event.params.tokenId).toHexString())
   );
-  if (!nft)
-    nft = new NFT(
-      Bytes.fromHexString(
-        event.params.nftContract
-          .toHexString()
-          .concat(Bytes.fromBigInt(event.params.tokenId).toHexString())
-      )
-    );
-
+  let nft = NFT.load(nftId);
+  if (!nft) nft = new NFT(nftId);
   // nft.seller = event.params.seller
-  // nft.category = event.params.category
+  // nft.category = event.params.category;
   nft.price = event.params.price;
   nft.tokenID = event.params.tokenId;
-  nft.owner = event.params.seller;
   nft.chain = 43113;
   nft.isListed = false;
   nft.isSold = true;
   nft.createdAtTimestamp = event.block.timestamp;
   nft.save();
+
+  let seller = User.load(event.params.seller);
+  if (!seller) seller = new User(event.params.seller);
+  let sellerNfts = seller.nfts;
+  if (!sellerNfts) sellerNfts = [];
+  let filteredNfts: Array<Bytes> = new Array<Bytes>();
+  for (let i = 0; i < sellerNfts.length; i++) {
+    const nft: Bytes = sellerNfts[i];
+    if (nft != nftId) {
+      filteredNfts.push(nft);
+    }
+  }
+  seller.nfts = filteredNfts;
+  seller.save();
 
   // let entity = new MarketItemSold(
   //   event.transaction.hash.concatI32(event.logIndex.toI32())
