+++ subgraph.yaml
@@ -3,75 +3,11 @@ schema:
   file: ./schema.graphql
 dataSources:
   - kind: ethereum
-    name: NftMinter
+    name: NftSoulBound
     network: fuji
     source:
-      address: "0xc37Da46d0ddF8F94428a273BCE25037bC54a885f"
-      abi: NftMinter
-      startBlock: 24529539
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.7
-      language: wasm/assemblyscript
-      entities:
-        - NFT
-        - Collection
-        - User
-      abis:
-        - name: NftMinter
-          file: ./abis/NftMinter.json
-      eventHandlers:
-        - event: ApprovalForAll(indexed address,indexed address,bool)
-          handler: handleApprovalForAll
-        - event: Initialized(uint8)
-          handler: handleInitialized
-        - event: TransferBatch(indexed address,indexed address,indexed address,uint256[],uint256[])
-          handler: handleTransferBatch1
-        - event: TransferSingle(indexed address,indexed address,indexed address,uint256,uint256)
-          handler: handleTransferSingle
-        - event: URI(string,indexed uint256)
-          handler: handleURI
-
-        - event: TransferBatch(indexed address,indexed address,indexed address,uint256[])
-          handler: handleTransferBatch
-      file: ./src/nft-minter.ts
-  - kind: ethereum
-    name: NftCollection
-    network: fuji
-    source:
-      address: "0xf73034a9f339FA6bad1e35CaC037605d767EdC0a"
-      abi: NftCollection
-      startBlock: 24529544
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.7
-      language: wasm/assemblyscript
-      entities:
-        - NFT
-        - Collection
-        - User
-      abis:
-        - name: NftCollection
-          file: ./abis/NftCollection.json
-      eventHandlers:
-        - event: AdminChanged(address,address)
-          handler: handleAdminChanged
-        - event: BeaconUpgraded(indexed address)
-          handler: handleBeaconUpgraded
-        - event: Initialized(uint8)
-          handler: handleInitialized
-        - event: Upgraded(indexed address)
-          handler: handleUpgraded
-
-        - event: CollectionCreated(indexed address,indexed address,string,string)
-          handler: handleCollectionCreated
-      file: ./src/nft-collection.ts
-  - kind: ethereum
-    name: SoulBound
-    network: fuji
-    source:
-      address: "0x0a06A3d6F23F1B3DC342c146DA0170D544c51532"
-      abi: SoulBound
+      address: "0x2998eCA937fad504a7Df30C8e1Ca403e4f16B036"
+      abi: NftSoulBound
       startBlock: 24677169
     mapping:
       kind: ethereum/events
@@ -82,8 +18,8 @@ dataSources:
         - Collection
         - User
       abis:
-        - name: SoulBound
-          file: ./abis/SoulBound.json
+        - name: NftSoulBound
+          file: ./abis/NftSoulBound.json
       eventHandlers:
         - event: Approval(indexed address,indexed address,indexed uint256)
           handler: handleApproval
@@ -91,13 +27,13 @@ dataSources:
           handler: handleApprovalForAll
         - event: Transfer(indexed address,indexed address,indexed uint256)
           handler: handleTransfer
-      file: ./src/soul-bound.ts
+      file: ./src/nft-soul-bound.ts
   - kind: ethereum
-    name: SingleNftMinter
+    name: NftSingle
     network: fuji
     source:
-      address: "0x41109163d8cf45E9dDd32801A8deb54b9513b1FA"
-      abi: SingleNftMinter
+      address: "0x862086adC9937E36e5a7989874D0A6BA0a153f8e"
+      abi: NftSingle
       startBlock: 24677157
     mapping:
       kind: ethereum/events
@@ -108,8 +44,8 @@ dataSources:
         - Collection
         - User
       abis:
-        - name: SingleNftMinter
-          file: ./abis/SingleNftMinter.json
+        - name: NftSingle
+          file: ./abis/NftSingle.json
       eventHandlers:
         - event: AdminChanged(address,address)
           handler: handleAdminChanged
@@ -132,12 +68,12 @@ dataSources:
           handler: handleCollectionInit
         - event: TransferBatch(indexed address,indexed address,indexed address,uint256[])
           handler: handleTransferBatch
-      file: ./src/single-nft-minter.ts
+      file: ./src/nft-single.ts
   - kind: ethereum
     name: NftMarket
     network: fuji
     source:
-      address: "0x53b922f0cA7B015bd71CaaCEd298b3a0E1710D08"
+      address: "0x078977d9Eb481Ed4de932c98bAC458f5c3c3705C"
       abi: NftMarket
       startBlock: 24678533
     mapping:
@@ -171,12 +107,12 @@ dataSources:
           handler: handleRecievedRoyalties
       file: ./src/nft-market.ts
   - kind: ethereum
-    name: NftMarketV2
+    name: NftFactory
     network: fuji
     source:
-      address: "0x420e727c31eD5D861fD27407DE6Be953d9Ca5bCE"
-      abi: NftMarketV2
-      startBlock: 24677192
+      address: "0x0e0878fbC369eEDb938885428c52465366206676"
+      abi: NftFactory
+      startBlock: 24529544
     mapping:
       kind: ethereum/events
       apiVersion: 0.0.7
@@ -186,45 +122,40 @@ dataSources:
         - Collection
         - User
       abis:
-        - name: NftMarketV2
-          file: ./abis/NftMarketV2.json
+        - name: NftFactory
+          file: ./abis/NftFactory.json
       eventHandlers:
         - event: AdminChanged(address,address)
           handler: handleAdminChanged
         - event: BeaconUpgraded(indexed address)
           handler: handleBeaconUpgraded
-        - event: BulkMarketItemCreated(indexed address,indexed address,string,uint256,uint256[],address,uint256,string)
-          handler: handleBulkMarketItemCreated
         - event: Initialized(uint8)
           handler: handleInitialized
-        - event: MarketItemCreated(indexed address,indexed address,string,uint256,uint256,address,uint256)
-          handler: handleMarketItemCreated
-        - event: MarketItemSold(indexed address,indexed address,uint256,uint256)
-          handler: handleMarketItemSold
-        - event: RecievedRoyalties(indexed address,indexed address,indexed uint256)
-          handler: handleRecievedRoyalties
         - event: Upgraded(indexed address)
           handler: handleUpgraded
-      file: ./src/nft-market-v2.ts
+
+        - event: CollectionCreated(indexed address,indexed address,string,string)
+          handler: handleCollectionCreated
+      file: ./src/nft-factory.ts
 
 templates:
   - kind: ethereum/contract
-    name: NftMinter
+    name: NftCollection
     network: fuji
     source:
-      abi: NftMinter
+      abi: NftCollection
     mapping:
       kind: ethereum/events
       apiVersion: 0.0.7
       language: wasm/assemblyscript
-      file: ./src/nft-minter-template.ts
+      file: ./src/nft-collection-template.ts
       entities:
         - NFT
         - Collection
         - User
       abis:
-        - name: NftMinter
-          file: ./abis/NftMinter.json
+        - name: NftCollection
+          file: ./abis/NftCollection.json
       eventHandlers:
         - event: ApprovalForAll(indexed address,indexed address,bool)
           handler: handleApprovalForAll
