+++ tests/nft-collection.test.ts
@@ -4,13 +4,13 @@ import {
   test,
   clearStore,
   beforeAll,
-  afterAll
-} from "matchstick-as/assembly/index"
-import { Address } from "@graphprotocol/graph-ts"
-import { ExampleEntity } from "../generated/schema"
-import { AdminChanged } from "../generated/NftCollection/NftCollection"
-import { handleAdminChanged } from "../src/nft-collection"
-import { createAdminChangedEvent } from "./nft-collection-utils"
+  afterAll,
+} from "matchstick-as/assembly/index";
+import { Address } from "@graphprotocol/graph-ts";
+import { ExampleEntity } from "../generated/schema";
+import { AdminChanged } from "../generated/NftCollection/NftCollection";
+import { handleAdminChanged } from "../src/nft-factory";
+import { createAdminChangedEvent } from "./nft-collection-utils";
 
 // Tests structure (matchstick-as >=0.5.0)
 // https://thegraph.com/docs/en/developer/matchstick/#tests-structure-0-5-0
@@ -19,23 +19,23 @@ describe("Describe entity assertions", () => {
   beforeAll(() => {
     let previousAdmin = Address.fromString(
       "0x0000000000000000000000000000000000000001"
-    )
+    );
     let newAdmin = Address.fromString(
       "0x0000000000000000000000000000000000000001"
-    )
-    let newAdminChangedEvent = createAdminChangedEvent(previousAdmin, newAdmin)
-    handleAdminChanged(newAdminChangedEvent)
-  })
+    );
+    let newAdminChangedEvent = createAdminChangedEvent(previousAdmin, newAdmin);
+    handleAdminChanged(newAdminChangedEvent);
+  });
 
   afterAll(() => {
-    clearStore()
-  })
+    clearStore();
+  });
 
   // For more test scenarios, see:
   // https://thegraph.com/docs/en/developer/matchstick/#write-a-unit-test
 
   test("ExampleEntity created and stored", () => {
-    assert.entityCount("ExampleEntity", 1)
+    assert.entityCount("ExampleEntity", 1);
 
     // 0xa16081f360e3847006db660bae1c6d1b2e17ec2a is the default address used in newMockEvent() function
     assert.fieldEquals(
@@ -43,15 +43,15 @@ describe("Describe entity assertions", () => {
       "0xa16081f360e3847006db660bae1c6d1b2e17ec2a",
       "previousAdmin",
       "0x0000000000000000000000000000000000000001"
-    )
+    );
     assert.fieldEquals(
       "ExampleEntity",
       "0xa16081f360e3847006db660bae1c6d1b2e17ec2a",
       "newAdmin",
       "0x0000000000000000000000000000000000000001"
-    )
+    );
 
     // More assert options:
     // https://thegraph.com/docs/en/developer/matchstick/#asserts
-  })
-})
+  });
+});
